backend/pyproject.toml
[project]
name = "drs-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg2-binary",
    "geoalchemy2",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "python-multipart",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "python-socketio",
    "python-engineio",
    "openai",
    "langchain",
    "pydantic-extra-types",
    "tenacity",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
]

[project.scripts]
app = "app.main:app"
